Send web request to openweather with required parameters.

+++ OpenWeatherClient/Abstractions/IWeather.cs
@@ -4,6 +4,6 @@ namespace OpenWeatherClient.Abstractions
 {
     public interface IWeather
     {
-         Task<decimal> CurrentTempAsync(Coord location);
+         Task<double> CurrentTempAsync(Coord location);
     }
 }
